% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunLoadDB.R
\name{calibrateRT}
\alias{calibrateRT}
\title{calibrateRT}
\usage{
calibrateRT(
  file_rt_ref = "RT_recalibration_table.csv",
  lib_rt,
  is_rt_calibration = TRUE,
  is_plot = TRUE,
  method_lc = c("Amide12min", "Amide23min"),
  column = "hilic",
  path = "."
)
}
\arguments{
\item{file_rt_ref}{Default: 'RT_recalibration_table.csv'}

\item{lib_rt}{a vector of RT values}

\item{is_rt_calibration}{whether need to calibrate RT}

\item{is_plot}{Default: TRUE}

\item{method_lc}{'Amide12min', 'Amide23min'. Default:  'Amide12min'}

\item{column}{Default: 'hilic'}

\item{path}{'.'}
}
\description{
RT calibration according to the RTQC
}
\author{
Zhiwei Zhou, Mingdu Luo
\email{zhouzw@sioc.ac.cn}
}
