% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunExportDB.R
\name{export_db_msdial}
\alias{export_db_msdial}
\title{export_db_msdial}
\usage{
export_db_msdial(
  lib = c("stanford", "iroa"),
  column = c("hilic", "c18"),
  polarity = c("positive", "negative"),
  msp_name = "dodd_lib_stanford_hilic_pos.msp",
  dir_path = "."
)
}
\arguments{
\item{lib}{dodd lab library. "stanford", "iroa"}

\item{column}{columne type. "hilic", "rp"}

\item{polarity}{ionization polarity. "positive", "negative"}

\item{msp_name}{ouput msp file name. e.g. "dodd_lib_stanford_hilic_pos.msp"}

\item{dir_path}{ouput directory path. Defaule: "."}
}
\description{
export_db_msdial
}
\examples{
\dontrun{
export_db_msdial(lib = 'stanford',
                 column = 'hilic',
                 polarity = 'negative',
                 msp_name = 'stanford_hilic_neg.msp',
                 dir_path = '~/Project/00_IBD_project/Data/20230215_MS2_lib_processing/04_final_tables')
}
}
\author{
Zhiwei Zhou
}
