% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunExportDB.R
\name{export_db_tidymass}
\alias{export_db_tidymass}
\title{export_db_tidymass}
\usage{
export_db_tidymass(
  lib = c("dodd", "stanford", "iroa"),
  column = c("hilic", "rp"),
  polarity = c("positive", "negative")
)
}
\arguments{
\item{lib}{dodd lab library. "stanford", "iroa"}

\item{column}{columne type. "hilic", "rp"}

\item{polarity}{ionization polarity. "positive", "negative"}
}
\description{
export_db_tidymass
}
\examples{
\dontrun{
stanford_hilic_pos_v1.0 <- export_db_tidymass(lib = 'stanford',
column = 'hilic',
polarity = 'positive')
}
}
\author{
Zhiwei Zhou
}
