% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunExportDB.R
\name{export_ms2_db}
\alias{export_ms2_db}
\title{export_ms2_db}
\usage{
export_ms2_db(
  lib = c("dodd"),
  ms1_version = c("v2.7.1", "v2.6.1", "v2.5.0", "v2.4.0", "v2.0.0"),
  ms2_version = c("v2.7.1", "v2.6.1", "v2.5.0", "v2.4.0", "v2.3.0"),
  polarity = c("positive", "negative"),
  msp_name = "dodd_lib_pos.msp",
  dir_path = "."
)
}
\arguments{
\item{lib}{dodd lab library. "dodd"}

\item{ms1_version}{MS1 version. "v2.6.1", "v2.5.0", "v2.4.0", "v2.0.0"}

\item{ms2_version}{MS2 version. "v2.6.1", "v2.5.0", "v2.4.0", "v2.3.0"}

\item{polarity}{ionization polarity. "positive", "negative"}

\item{msp_name}{ouput msp file name. e.g. "dodd_lib_stanford_hilic_pos.msp"}

\item{dir_path}{ouput directory path. Defaule: "."}
}
\description{
export_ms2_db
}
\examples{
\dontrun{
export_ms2_db(lib = 'dodd',
              ms1_version = 'v2.6.1',
              ms2_version = 'v2.6.1',
              polarity = 'positive',
              msp_name = "dodd_lib_pos_v2.6.1.msp",
              dir_path = '~/Project/04_package/00_Database/DoddLib/06_exported_db/')

export_ms2_db(lib = 'dodd',
              ms1_version = 'v2.6.1',
              ms2_version = 'v2.6.1',
              polarity = 'negative',
              msp_name = "dodd_lib_neg_v2.6.1.msp",
              dir_path = '~/Project/04_package/00_Database/DoddLib/06_exported_db/')
}
}
\author{
Zhiwei Zhou
}
