% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunBuildDB.R
\name{extract_ms2_from_target_list}
\alias{extract_ms2_from_target_list}
\title{extract_ms2_from_target_list}
\usage{
extract_ms2_from_target_list(
  target_list_file,
  dir_path,
  polarity = c("positive", "negative")
)
}
\arguments{
\item{target_list_file}{a target list file. The target list file should include id, mz, rt}

\item{dir_path}{the directory path of the mzml files}

\item{polarity}{the polarity of the MS2 data, 'positive' or 'negative'. Default: 'positive'}
}
\description{
extract ms2 from target list and exported into MSP files
}
\examples{
\dontrun{
extract_ms2_from_target_list(target_list_file = 'Targeted_MS2_List.xlsx',
                             dir_path = '~/Project/05_coworker/Kazuma/241014/demo_data',
                             polarity = 'positive')
}
}
\author{
Zhiwei Zhou
}
