% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunExportDB.R
\name{generate_msp}
\alias{generate_msp}
\title{generate_msp}
\usage{
generate_msp(
  file_name = "./test.msp",
  cmp_name = NULL,
  precusormz = NULL,
  adduct = NULL,
  instrument_type = NULL,
  instrument = NULL,
  smiles = NULL,
  inchikey = NULL,
  inchikey1 = NULL,
  formula = NULL,
  polarity = c("positive", "negative"),
  ce = NULL,
  rt = NULL,
  ccs = NULL,
  doddlib_id = NULL,
  kegg_id = NULL,
  hmdb_id = NULL,
  pubchem_id = NULL,
  links = "",
  comment = "",
  spec = NULL
)
}
\arguments{
\item{file_name}{Required. The file name of MSP. The suffix of file name must be ".msp".}

\item{cmp_name}{Required}

\item{precusormz}{Required}

\item{adduct}{Default: NULL}

\item{instrument_type}{Default: NULL}

\item{instrument}{Default: NULL}

\item{smiles}{Default: NULL}

\item{inchikey}{Default: NULL}

\item{inchikey1}{Default: NULL}

\item{formula}{Default: NULL}

\item{polarity}{'Positive' or 'Negative'}

\item{ce}{Default: NULL}

\item{rt}{Default: NULL}

\item{ccs}{Default: NULL}

\item{doddlib_id}{Default: NULL}

\item{kegg_id}{Default: NULL}

\item{hmdb_id}{Default: NULL}

\item{pubchem_id}{Default: NULL}

\item{links}{Default: ''}

\item{comment}{Default: ''}

\item{spec}{Required. It should be in a dataframe, 1st: "mz", 2nd: "intensity"}
}
\description{
Convert MS/MS spectra to MSP files
}
\examples{
setwd('F:/01 MetIMMS/00 data processing/190515 external validation msms data extraction/')
    GenerateMSP(file_name = 'zhumetlib_validation_pos_20v_190520.msp',
                cmp_name = external_data_pos$compound_name[i],
                precusormz = external_data_pos$mz[i],
                adduct = external_data_pos$adducts[i],
                instrument_type = 'LC-ESI-qTof',
                instrument = 'LC-ESI-qTof',
                smiles = external_data_pos$smiles[i],
                inchikey = external_data_pos$inchikey[i],
                inchikey1 = external_data_pos$inchikey1[i],
                formula = external_data_pos$formula[i],
                polarity = 'Positive',
                ce = '20',
                ccs = external_data_pos$CCS[i],
                zhulib_id = external_data_pos$matched_zhulib_id[i],
                pubchem_id = external_data_pos$pubchem_cid[i],
                comment = paste(external_data_pos$id[i], external_data_pos$source[i], sep = ' '),
                spec = temp_spec)
}
\author{
Zhiwei Zhou
}
