% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunBuildDB.R
\name{group_ms2}
\alias{group_ms2}
\title{group_ms2}
\usage{
group_ms2(
  target_list,
  ms2_data,
  label = "",
  mode = "hilic_pos",
  mz_tol = 15,
  rt_tol = 15,
  rt_error_type = "abs"
)
}
\arguments{
\item{target_list}{a target list (data.frame). First 3 columns, id, mz, rt}

\item{ms2_data}{ms2 data. Generated by retrieve_ms2}

\item{label}{Default: ''. Usually can label batch name, e.g. '2023_Zhiwei'}

\item{mode}{Default: ''. Usually can label column and polarity, e.g. 'hilic_pos'}

\item{mz_tol}{Default: 15. Match target list with ms2_info tolerance}

\item{rt_tol}{Default: 15. Match target list with ms2_info tolerance}
}
\description{
group_ms2
}
\examples{
\dontrun{
}
}
\author{
Zhiwei Zhou
}
