% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunBuildDB.R
\name{merge_ms2}
\alias{merge_ms2}
\title{merge_ms2}
\usage{
merge_ms2(ms2_list)
}
\arguments{
\item{ms2_list}{a list of STD ms2 data.}
}
\description{
merge the ms2 list, and organized with ID
}
\examples{
\dontrun{
load('~/Project/00_IBD_project/Data/20230428_Zhiwei_StanfordPool_MS2/mzxml_haoqing_data/Haoqing_MS2_data/MS2_lib_230509/ms2_stanford_C18_pos_raw_230509.RData')
load('~/Project/00_IBD_project/Data/20230428_Zhiwei_StanfordPool_MS2/mzxml_haoqing_data/Haoqing_MS2_data/MS2_lib_230509/ms2_stanford_HILIC_pos_raw_230509.RData')

# combine these two ms2 list as one list
ms2_list <- c(ms2_stanford_HILIC_pos_raw, ms2_stanford_C18_pos_raw)
ms2_list <- merge_ms2(ms2_list)
}
}
\author{
Zhiwei Zhou
}
