% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunBuildDB.R
\name{select_best_ms2}
\alias{select_best_ms2}
\title{select_best_ms2}
\usage{
select_best_ms2(
  ms2_list,
  precursor_mz_error_cutoff = 0.0015,
  precursor_rt_error_cutoff = 15,
  precursor_int_cutoff = 500,
  precursor_purity_cutoff = 0.9,
  is_keep_conflict_std = TRUE,
  rank_by = c("intensity", "intensity_purity", "rt_error")
)
}
\arguments{
\item{ms2_list}{a merged ms2 list. The merged ms2 list can be generated by merge_ms2()}

\item{precursor_mz_error_cutoff}{the precursor mz error cutoff. Default: 0.0015 Da}

\item{precursor_rt_error_cutoff}{the precursor rt error cutoff. Default: 15 second}

\item{precursor_int_cutoff}{the minimum precursor intensity. Default: 500 counts}

\item{precursor_purity_cutoff}{the minmium precursor purity. Default: 0.9}

\item{is_keep_conflict_std}{whether keep the possible conflict ms2 spec. e.g. some targeted isobaric/isomers run in one injection, and have very similar RT. Default: TRUE}

\item{rank_by}{the way to select the best spec, includes "intensity", "intensity_purity", "rt_error". Default: "intensity"}
}
\description{
select_best_ms2
}
\examples{
\dontrun{
load('~/Project/00_IBD_project/Data/20230428_Zhiwei_StanfordPool_MS2/mzxml_haoqing_data/Haoqing_MS2_data/MS2_lib_230509/ms2_stanford_C18_pos_raw_230509.RData')
load('~/Project/00_IBD_project/Data/20230428_Zhiwei_StanfordPool_MS2/mzxml_haoqing_data/Haoqing_MS2_data/MS2_lib_230509/ms2_stanford_HILIC_pos_raw_230509.RData')

# combine these two ms2 list as one list
ms2_list <- c(ms2_stanford_HILIC_pos_raw, ms2_stanford_C18_pos_raw)
ms2_list <- merge_ms2(ms2_list)
select_best_ms2(ms2_list = ms2_list,
                precursor_mz_error_cutoff = 0.0015, # absolute mz error, Da
                precursor_rt_error_cutoff = 15, # absolute rt error, second
                precursor_int_cutoff = 500,
                precursor_purity_cutoff = 0.9,
                is_keep_conflict_std = TRUE,
                rank_by = "intensity_purity")
}
}
\author{
Zhiwei Zhou
}
