% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunSearchDB.R
\name{state_library}
\alias{state_library}
\title{state_library}
\usage{
state_library()
}
\description{
show the statistics of all libraries
}
\examples{
\dontrun{
state_library()
}
}
\author{
Zhiwei Zhou
}
