% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunLoadDB.R
\name{load_spec_db}
\alias{load_spec_db}
\title{load_spec_db}
\usage{
load_spec_db(
  lib = c("dodd", "msdial", "peptide", "gnps_bile_acid", "gnps_acyl_amides",
    "gnps_acyl_esters", "all_public"),
  column = c("hilic", "c18"),
  ce = c("10", "20", "40"),
  polarity = c("positive", "negative"),
  adduct_list = NULL
)
}
\arguments{
\item{lib}{database name, 'dodd', 'msdial', 'peptide'. Default: 'dodd'}

\item{column}{'hilic', 'c18'. Default: 'hilic'}

\item{ce}{"10", "20", "40"; Default: '20'}

\item{polarity}{'positive' or 'negative'. Default: 'positive'}

\item{adduct_list}{NULL}
}
\description{
load_spec_db
}
\examples{
\dontrun{
# dodd lib
test <- load_spec_db(lib = 'dodd', column = 'hilic', ce = '20', polarity = 'positive', adduct_list = '[M+H]+')
# msdial lib
test <- load_spec_db(lib = 'msdial', polarity = 'positive')
# peptide lib
test <- load_spec_db(lib = 'peptide', polarity = 'positive')
}

}
\author{
Zhiwei Zhou
}
