% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunLoadDB.R
\name{load_spec_db}
\alias{load_spec_db}
\title{load_spec_db}
\usage{
load_spec_db(
  lib = c("dodd", "msdial"),
  column = c("hilic", "c18"),
  ce = c("10", "20", "40"),
  polarity = c("positive", "negative"),
  adduct_list = NULL
)
}
\arguments{
\item{lib}{database name, 'zhuMetLib', 'zhuMetLib_orbitrap', 'fiehnHilicLib'. Default: 'zhuMetLib'}

\item{column}{'hilic', 'rp'. Default: 'hilic'}

\item{ce}{"10", "20", "30", "35,15", "40", "50"; Default: '30'}

\item{polarity}{'positive' or 'negative'. Default: 'positive'}

\item{adduct_list}{NULL}

\item{instrument}{"SciexTripleTOF", "AgilentQTOF", "BrukerQTOF", "ThermoOrbitrap", "ThermoExploris", "AgilentDTIMMS", "BrukerTIMS". Default: "SciexTripleTOF"}

\item{method_lc}{'Amide12min' or 'Amide23min'}

\item{is_rt_calibration}{TRUE}

\item{path}{'.'}
}
\description{
load_spec_db
}
\examples{
\dontrun{
# dodd lib
test <- load_spec_db(lib = 'dodd', column = 'hilic', ce = '20', polarity = 'positive', adduct_list = '[M+H]+')
# msdial lib
test <- load_spec_db(lib = 'msdial', polarity = 'positive')
}

}
\author{
Zhiwei Zhou
}
